@onur-ozkan 2024-01-18
> I think this just broke building everything with -Zthreads=8 - https://internals.rust-lang.org/t/rust-compiler-with-parallel-build/20099

Correct.

> If RUSTC_BOOTSTRAP was set externally, then it should not be removed.

I think setting it consistently is much better approach. What do you think ? @dtolnay