@dtolnay 2024-01-18
> I think setting it consistently is much better approach (or if it wasn't manually set from user). What do you think ? @dtolnay

I agree with this, but responded in https://github.com/rust-lang/rust/pull/120096#pullrequestreview-1830037089 with one modification I would suggest.